Exception handling - update

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
@@ -33,9 +32,10 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        try{
+        try {
+            if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
             repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
+        } catch (ResourceNotFoundException e) {
             throw new ResourceNotFoundException(id);
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException(e.getMessage());
@@ -43,9 +43,13 @@ public class UserService {
     }
 
     public User update(Long id, User obj) {
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        } catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
